+++ src/main/java/CreateOrderPOJO.java
@@ -1,2 +1,77 @@
-package PACKAGE_NAME;public class CreateOrderPOJO {
+import java.util.ArrayList;
+
+public class CreateOrderPOJO {
+    private String firstName;
+    private String lastName;
+    private String address;
+    private String metroStation;
+    private String phone;
+    private int rentTime;
+    private String deliveryDate;
+    private String comment;
+    private String[] color;
+
+    public CreateOrderPOJO(String firstName, String lastName, String address, String metroStation, String phone, int rentTime, String deliveryDate, String comment, String[] color) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.address = address;
+        this.metroStation = metroStation;
+        this.phone = phone;
+        this.rentTime = rentTime;
+        this.deliveryDate = deliveryDate;
+        this.comment = comment;
+        this.color = color;
+    }
+
+    public CreateOrderPOJO(String firstName, String lastName, String address, String metroStation, String phone, int rentTime, String deliveryDate, String comment) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.address = address;
+        this.metroStation = metroStation;
+        this.phone = phone;
+        this.rentTime = rentTime;
+        this.deliveryDate = deliveryDate;
+        this.comment = comment;
+    }
+
+    public CreateOrderPOJO() {
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public String getMetroStation() {
+        return metroStation;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public int getRentTime() {
+        return rentTime;
+    }
+
+    public String getDeliveryDate() {
+        return deliveryDate;
+    }
+
+    public String getComment() {
+        return comment;
+    }
+
+    public String[] getColor() {
+        return color;
+    }
+
+
 }
