+++ src/test/java/ScooterDeleteCourierTest.java
@@ -3,6 +3,7 @@ import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.RandomUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -14,12 +15,11 @@ import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
 
-public class ScooterLoginCourierTest {
+public class ScooterDeleteCourierTest {
 
 
-    private String login;
-    private String password;
-    private String random = RandomStringUtils.randomAlphabetic(10);
+    private int randomNotExist = RandomUtils.nextInt(10000000, 99999999);
+    private String courierId;
 
 
     @Before
@@ -27,13 +27,7 @@ public class ScooterLoginCourierTest {
         RestAssured.baseURI = "https://qa-scooter.praktikum-services.ru/";
         ScooterRegisterCourier scooterRegisterCourier = new ScooterRegisterCourier();
         ArrayList<String> loginPass = scooterRegisterCourier.registerNewCourierAndReturnLoginPassword();
-        login = loginPass.get(0);
-        password = loginPass.get(1);
-    }
-
-    @After
-    public void tearDown() {
-        RegisterCourierPOJO registerCourierLogin = new RegisterCourierPOJO(login, password);
+        RegisterCourierPOJO registerCourierLogin = new RegisterCourierPOJO(loginPass.get(0), loginPass.get(1));
         String response = given()
                 .header("Content-type", "application/json")
                 .and()
@@ -42,189 +36,50 @@ public class ScooterLoginCourierTest {
                 .post("/api/v1/courier/login")
                 .asString();
         JsonPath jsonPath = new JsonPath(response);
-        String userId = jsonPath.getString("id");
-        delete("/api/v1/courier/" + userId);
+        courierId = jsonPath.getString("id");
+    }
+
+    @After
+    public void tearDown() {
+        delete("/api/v1/courier/" + courierId);
     }
 
     @Test
-    @DisplayName("Check status code of /api/v1/courier/login when data is valid")
-    public void checkStatusCodeLoginCourierWithValidData() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(login, password);
+    @DisplayName("Check message error of /api/v1/courier/:id when id is non-exist")
+    public void checkMessageErrorDeleteCourierWhenIdNonExist() {
         Response response = given()
                 .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
                 .when()
-                .post("/api/v1/courier/login");
+                .delete("/api/v1/courier/" + randomNotExist);
         response.then()
                 .assertThat()
-                .statusCode(200);
+                .body("message", equalTo("Курьера с таким id нет"));
     }
 
     @Test
-    @DisplayName("Check id of /api/v1/courier/login when data is valid")
-    public void checkIdAfterLoginCourierWithValidData() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(login, password);
+    @DisplayName("Check body of /api/v1/courier/:id when id is valid")
+    public void checkMessageDeleteCourierWhenDataIsValid() {
         Response response = given()
                 .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
                 .when()
-                .post("/api/v1/courier/login");
+                .delete("/api/v1/courier/" + courierId);
         response.then()
                 .assertThat()
-                .body("id", notNullValue());
+                .body("ok", equalTo(true));
     }
 
     @Test
-    @DisplayName("Check status code of /api/v1/courier/login when login is valid, but password is invalid")
-    public void checkStatusCodeLoginCourierWithIncorrectPassword() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(login, random);
+    @DisplayName("Check message error of /api/v1/courier/:id when id is missing")
+    public void checkMessageErrorDeleteCourierWhenIdMissing() {
         Response response = given()
                 .header("Content-type", "application/json")
                 .and()
-                .body(registerCourierPOJO)
+                .body("{}")
                 .when()
-                .post("/api/v1/courier/login");
+                .delete("/api/v1/courier/");
         response.then()
                 .assertThat()
-                .statusCode(404);
-    }
-
-    @Test
-    @DisplayName("Check message of /api/v1/courier/login when login is valid, but password is invalid")
-    public void checkMessageLoginCourierWithIncorrectPassword() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(login, random);
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .body("message", equalTo("Учетная запись не найдена"));
-    }
-
-    @Test
-    @DisplayName("Check status code of /api/v1/courier/login when login is invalid, but password is valid")
-    public void checkStatusCodeLoginCourierWithIncorrectLogin() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(random, password);
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .statusCode(404);
-    }
-
-    @Test
-    @DisplayName("Check message of /api/v1/courier/login when login is invalid, but password is valid")
-    public void checkMessageLoginCourierWithIncorrectLogin() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(random, password);
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .body("message", equalTo("Учетная запись не найдена"));
-    }
-
-    @Test
-    @DisplayName("Check status code of /api/v1/courier/login when login is invalid, but password is invalid")
-    public void checkStatusCodeLoginWithNonExistCourier() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(random, random);
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .statusCode(404);
-    }
-
-    @Test
-    @DisplayName("Check message of /api/v1/courier/login when login is invalid, but password is invalid")
-    public void checkMessageLoginWithNonExistCourier() {
-        RegisterCourierPOJO registerCourierPOJO = new RegisterCourierPOJO(random, random);
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerCourierPOJO)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .body("message", equalTo("Учетная запись не найдена"));
-    }
-
-
-    @Test
-    @DisplayName("Check status code of /api/v1/courier/login when data is invalid, missing password field")
-    public  void checkStatusCodeLoginCourierWithoutFillInPassword() {
-        String registerBody = "{\"login\":\"" + login + "\"}";
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerBody)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("Check message of /api/v1/courier/login when data is invalid, missing password field")
-    public  void checkMessageLoginCourierWithoutFillInPassword() {
-        String registerBody = "{\"login\":\"" + login + "\"}";
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerBody)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .body("message", equalTo("Недостаточно данных для входа"));
-    }
-
-    @Test
-    @DisplayName("Check status code of /api/v1/courier/login when data is invalid, missing login field")
-    public  void checkStatusCodeLoginCourierWithoutFillInLogin() {
-        String registerBody = "{\"password\":\"" + password + "\"}";
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerBody)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("Check message of /api/v1/courier/login when data is invalid, missing login field")
-    public  void checkMessageLoginCourierWithoutFillInLogin() {
-        String registerBody = "{\"password\":\"" + password + "\"}";
-        Response response = given()
-                .header("Content-type", "application/json")
-                .and()
-                .body(registerBody)
-                .when()
-                .post("/api/v1/courier/login");
-        response.then()
-                .assertThat()
-                .body("message", equalTo("Недостаточно данных для входа"));
+                .body("message", equalTo("Недостаточно данных для удаления курьера"));
     }
 
 }
