+++ src/test/java/ScooterGetOrderByTrackTest.java
@@ -2,6 +2,7 @@ import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.junit.After;
 import org.junit.Before;
@@ -9,69 +10,71 @@ import org.junit.Test;
 
 import java.util.ArrayList;
 
-import static io.restassured.RestAssured.*;
+import static io.restassured.RestAssured.delete;
+import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.notNullValue;
 
-public class ScooterApplyOrderByCourierTest {
+public class ScooterGetOrderByTrackTest {
 
 
     private int randomNotExist = RandomUtils.nextInt(10000000, 99999999);
-    private int courierId;
-    private int orderId;
+    private int trackId;
+    static private String randomString = RandomStringUtils.randomAlphabetic(10);
+    static private int randomInt = RandomUtils.nextInt(1, 11);
+    static private String randomDate = "2020-06-" + RandomUtils.nextInt(11, 30);
 
 
     @Before
     public void setUp() {
         RestAssured.baseURI = "https://qa-scooter.praktikum-services.ru/";
-        ScooterRegisterCourier scooterRegisterCourier = new ScooterRegisterCourier();
-        ArrayList<String> loginPass = scooterRegisterCourier.registerNewCourierAndReturnLoginPassword();
-        RegisterCourierPOJO registerCourierLogin = new RegisterCourierPOJO(loginPass.get(0), loginPass.get(1));
-        String response = given()
+        CreateOrderPOJO createOrderPOJO = new CreateOrderPOJO(randomString, randomString, randomString, randomString,
+                randomString, randomInt, randomDate, randomString);
+        Response response = given()
                 .header("Content-type", "application/json")
                 .and()
-                .body(registerCourierLogin)
+                .body(createOrderPOJO)
                 .when()
-                .post("/api/v1/courier/login")
-                .asString();
-        JsonPath jsonPath = new JsonPath(response);
-        courierId = jsonPath.getInt("id");
-        String responseOrders = given()
-                .header("Content-type", "application/json")
-                .when()
-                .get("/api/v1/orders")
-                .asString();
-        JsonPath jsonPathOfOrderId = new JsonPath(responseOrders);
-        orderId = jsonPathOfOrderId.getInt("orders[0].id");
+                .post("/api/v1/orders");
+        String responseString = response.asString();
+        JsonPath jsonPath = new JsonPath(responseString);
+        trackId = jsonPath.getInt("track");
     }
 
     @After
     public void tearDown() {
-        delete("/api/v1/courier/" + courierId);
+        String cancelBody = "{\"track\":" + trackId + "}";
+        given()
+                .header("Content-type", "application/json")
+                .and()
+                .body(cancelBody)
+                .when()
+                .put("/api/v1/orders/cancel");
     }
 
     @Test
-    @DisplayName("Check success body of /api/v1/courier/accept/:id?courierId")
-    public void checkSuccessBodyApplyByCourierWhenDataIsValid() {
+    @DisplayName("Check success body of /api/v1/orders/track/t?")
+    public void checkSuccessBodyGetByTrackOrderListWhenDataIsValid() {
         Response response = given()
                 .header("Content-type", "application/json")
                 .when()
-                .queryParam("courierId", courierId)
-                    .put("api/v1/orders/accept/" + orderId);
+                .queryParam("t", trackId)
+                    .get("api/v1/orders/track");
         response.then()
                 .assertThat()
                 .statusCode(200)
                 .and()
-                .body("ok", equalTo(true));
+                .body("order", notNullValue());
     }
 
     @Test
-    @DisplayName("Check error body of /api/v1/courier/accept/:id?courierId when orderID missing")
-    public void checkErrorBodyApplyByCourierWhenOrderIdMissing() {
+    @DisplayName("Check error body of /api/v1/orders/track/t? when track is missing")
+    public void checkErrorBodyGetByTrackOrderListWhenTrackMissing() {
         Response response = given()
                 .header("Content-type", "application/json")
                 .when()
-                .queryParam("courierId", courierId)
-                .put("api/v1/orders/accept/");
+                .queryParam("t")
+                .get("api/v1/orders/track");
         response.then()
                 .assertThat()
                 .statusCode(400)
@@ -80,50 +83,18 @@ public class ScooterApplyOrderByCourierTest {
     }
 
     @Test
-    @DisplayName("Check error body of /api/v1/courier/accept/:id?courierId when courierId missing")
-    public void checkErrorBodyApplyByCourierWhenCourierIdMissing() {
+    @DisplayName("Check error body of /api/v1/orders/track/t? when track is non-exist")
+    public void checkErrorBodyGetByTrackOrderListWhenTrackNonExist() {
         Response response = given()
                 .header("Content-type", "application/json")
                 .when()
-                .queryParam("courierId")
-                .put("api/v1/orders/accept/" + orderId);
-        response.then()
-                .assertThat()
-                .statusCode(400)
-                .and()
-                .body("message", equalTo("Недостаточно данных для поиска"));
-    }
-
-    @Test
-    @DisplayName("Check error body of /api/v1/courier/accept/:id?courierId when courierId non-exist")
-    public void checkErrorBodyApplyByCourierWhenCourierIdNonExist() {
-        Response response = given()
-                .header("Content-type", "application/json")
-                .when()
-                .queryParam("courierId", randomNotExist)
-                .put("api/v1/orders/accept/" + orderId);
+                .queryParam("t", randomNotExist)
+                .get("api/v1/orders/track");
         response.then()
                 .assertThat()
                 .statusCode(404)
                 .and()
-                .body("message", equalTo("Курьера с таким id не существует"));
+                .body("message", equalTo("Заказ не найден"));
     }
-
-    @Test
-    @DisplayName("Check error body of /api/v1/courier/accept/:id?courierId when orderId non-exist")
-    public void checkErrorBodyApplyByCourierWhenOrderIdNonExist() {
-        Response response = given()
-                .header("Content-type", "application/json")
-                .when()
-                .queryParam("courierId", courierId)
-                .put("api/v1/orders/accept/" + randomNotExist);
-        response.then()
-                .assertThat()
-                .statusCode(404)
-                .and()
-                .body("message", equalTo("Заказа с таким id не существует"));
-    }
-
-
 
 }
